backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pytest-postgresql (>=6.1.1,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "redis (>=5.2.1,<6.0.0)",
    "python-jose (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
